Removed table and user DROPs, because database is newly created and added new table creations

+++ setup/index.php
@@ -68,30 +68,31 @@ switch (empty($_GET['setup']) ? '' : $_GET['setup']) {
 				echo '<p>Database clans_of_macaria selected successfully</p>';
 
 				// Tables
-				$sql = 'DROP TABLE games';
-				if (mysqli_query($con, $sql))
-					echo '<p>Table games deleted successfully</p>';
-				else
-					echo '<p>Error deleting table games: ' . mysqli_error($con) . '</p>';
-
-				$sql = 'CREATE TABLE games (gameid BIGINT NOT NULL AUTO_INCREMENT, name VARCHAR(75), PRIMARY KEY (gameid))';
+				$sql = 'CREATE TABLE games (gameid BIGINT NOT NULL AUTO_INCREMENT, name VARCHAR(75) UNIQUE NOT NULL, PRIMARY KEY (gameid))';
 				if (mysqli_query($con, $sql))
 					echo '<p>Table games created successfully</p>';
 				else
-					echo '<p>Error creating table: ' . mysqli_error($con) . '</p>';
+					echo '<p>Error creating table games: ' . mysqli_error($con) . '</p>';
 
-				$sql = 'DROP TABLE players';
-				if (mysqli_query($con, $sql))
-					echo '<p>Table players deleted successfully</p>';
-				else
-					echo '<p>Error deleting table players: ' . mysqli_error($con) . '</p>';
-
-				$sql = 'CREATE TABLE players (playerid BIGINT NOT NULL AUTO_INCREMENT, name VARCHAR(75), PRIMARY KEY (playerid))';
+				$sql = 'CREATE TABLE players (playerid BIGINT NOT NULL AUTO_INCREMENT, name VARCHAR(75) NOT NULL, username VARCHAR(25) UNIQUE NOT NULL, password CHAR(18) NOT NULL, PRIMARY KEY (playerid))';
 				if (mysqli_query($con, $sql))
 					echo '<p>Table players created successfully</p>';
 				else
 					echo '<p>Error creating table players: ' . mysqli_error($con) . '</p>';
 
+				$sql = 'CREATE TABLE clan_types (clan_typeid BIGINT NOT NULL AUTO_INCREMENT, name VARCHAR(75) UNIQUE NOT NULL, image_url VARCHAR(255) NOT NULL, PRIMARY KEY (clan_typeid))';
+				if (mysqli_query($con, $sql))
+					echo '<p>Table clan_types created successfully</p>';
+				else
+					echo '<p>Error creating table clan_types: ' . mysqli_error($con) . '</p>';
+
+				$sql = 'CREATE TABLE games_to_players_to_clan_types (gameid BIGINT NOT NULL, playerid BIGINT NOT NULL, clan_typeid BIGINT NOT NULL, FOREIGN KEY (gameid) REFERENCES games(gameid), FOREIGN KEY (playerid) REFERENCES players(playerid), FOREIGN KEY (clan_typeid) REFERENCES clan_types(clan_typeid))';
+				if (mysqli_query($con, $sql))
+					echo '<p>Table games_to_players_to_clan_types created successfully</p>';
+				else
+					echo '<p>Error creating table games_to_players_to_clan_types: ' . mysqli_error($con) . '</p>';
+
+				// Summarize what we just created
 				$sql = 'SHOW TABLES';
 				$result = mysqli_query($con, $sql);
 				if ($result) {
@@ -112,48 +113,43 @@ switch (empty($_GET['setup']) ? '' : $_GET['setup']) {
 				}
 				else
 					echo '<p>Error listing tables: ' . mysqli_error($con) . '</p>';
+
+				// Write out clans database user's password
+				if (file_put_contents('access.php', '<' . '?php $clans_of_macaria_password = \'' . str_replace('\'', "\\'", $_GET['clans']) . '\'; ?' . '>') > 0)
+					echo '<p>User clans password saved to ' . getcwd() . '/access.php</p>';
+				else
+					echo '<p>Error writing clans password</p>';
+
+				// Create clans database user
+				$sql = 'CREATE USER clans@clans_of_macaria IDENTIFIED BY \'' . str_replace('\'', "\\'", $_GET['clans']) . '\'';
+				if (mysqli_query($con, $sql))
+					echo '<p>User clans created successfully</p>';
+				else
+					echo '<p>Error creating user clans: ' . mysqli_error($con) . '</p>';
+
+				$sql = 'GRANT SELECT, INSERT, UPDATE, DELETE ON clans_of_macaria.* TO clans';
+				if (mysqli_query($con, $sql))
+					echo '<p>User clans given SELECT, INSERT, UPDATE and DELETE access to database clans_of_macaria successfully</p>';
+				else
+					echo '<p>Error giving full access: ' . mysqli_error($con) . '</p>';
+
+				$sql = 'FLUSH PRIVILEGES';
+				if (mysqli_query($con, $sql))
+					echo '<p>Flushed privileges successfully</p>';
+				else
+					echo '<p>Error flushing privileges: ' . mysqli_error($con) . '</p>';
+
+				$conclans = mysqli_connect('localhost', 'clans', $_GET['clans']);
+				if (mysqli_connect_errno())
+					echo '<p>Failed to connect to MySQL: ' . mysqli_connect_error() . '</p>';
+				else {
+					echo '<p>User clans connected successfully</p>';
+	
+					mysqli_close($conclans);
+				}
 			}
 			else
 				echo '<p>Error selecting database clans_of_macaria: ' . mysqli_error($con) . '</p>';
-
-			// Write out clans database user's password
-			file_put_contents('access.php', '<' . '?php $clans_of_macaria_password = \'' . str_replace('\'', "\\'", $_GET['clans']) . '\'; ?' . '>');
-			echo '<p>User clans password saved to ' . getcwd() . '/access.php</p>';
-
-			// Clans database user
-			$sql = 'DROP USER clans';
-			if (mysqli_query($con, $sql))
-				echo '<p>User clans deleted successfully</p>';
-			else
-				echo '<p>Error deleting user clans: ' . mysqli_error($con) . '</p>';
-
-			$sql = 'CREATE USER clans IDENTIFIED BY \'' . str_replace('\'', "\\'", $_GET['clans']) . '\'';
-			if (mysqli_query($con, $sql))
-				echo '<p>User clans created successfully</p>';
-			else
-				echo '<p>Error creating user clans: ' . mysqli_error($con) . '</p>';
-
-			$sql = 'GRANT SELECT, INSERT, UPDATE, DELETE ON clans_of_macaria.* TO clans';
-			if (mysqli_query($con, $sql))
-				echo '<p>User clans given SELECT, INSERT, UPDATE and DELETE access to database clans_of_macaria successfully</p>';
-			else
-				echo '<p>Error giving full access: ' . mysqli_error($con) . '</p>';
-
-			$sql = 'FLUSH PRIVILEGES';
-			if (mysqli_query($con, $sql))
-				echo '<p>Flushed privileges successfully</p>';
-			else
-				echo '<p>Error flushing privileges: ' . mysqli_error($con) . '</p>';
-
-			$conclans = mysqli_connect('localhost', 'clans', $_GET['clans']);
-			if (mysqli_connect_errno())
-				echo '<p>Failed to connect to MySQL: ' . mysqli_connect_error() . '</p>';
-			else {
-				echo '<p>User clans connected successfully</p>';
-
-				mysqli_close($conclans);
-			}
-
 		}
 
 		mysqli_close($con);
